Main, Admin, User (Student), and New User files
Main and New User files are complete, need to work on Admin and User files. Refer to the project report to build the framework for the code. Highly recommend you read through the files to understand my thought process and how I have divided the code. Don't worry about redundancy now, we can deal with that later.

Note that based on our implementation, I needed to make UID as a student's username (auto generated and assigned to a new student). Username and Password are new attributes to the Student and Admin tables.